add tests and move to beam.py
